+++ i18n/keys.go
@@ -237,17 +237,20 @@ const (
 
 // Setting related messages
 const (
-	MsgSettingInvalidType      = "setting.invalid_type"
-	MsgSettingWebhookEmpty     = "setting.webhook_empty"
-	MsgSettingWebhookInvalid   = "setting.webhook_invalid"
-	MsgSettingEmailInvalid     = "setting.email_invalid"
-	MsgSettingBarkUrlEmpty     = "setting.bark_url_empty"
-	MsgSettingBarkUrlInvalid   = "setting.bark_url_invalid"
-	MsgSettingGotifyUrlEmpty   = "setting.gotify_url_empty"
-	MsgSettingGotifyTokenEmpty = "setting.gotify_token_empty"
-	MsgSettingGotifyUrlInvalid = "setting.gotify_url_invalid"
-	MsgSettingUrlMustHttp      = "setting.url_must_http"
-	MsgSettingSaved            = "setting.saved"
+	MsgSettingInvalidType           = "setting.invalid_type"
+	MsgSettingWebhookEmpty          = "setting.webhook_empty"
+	MsgSettingWebhookInvalid        = "setting.webhook_invalid"
+	MsgSettingEmailInvalid          = "setting.email_invalid"
+	MsgSettingBarkUrlEmpty          = "setting.bark_url_empty"
+	MsgSettingBarkUrlInvalid        = "setting.bark_url_invalid"
+	MsgSettingGotifyUrlEmpty        = "setting.gotify_url_empty"
+	MsgSettingGotifyTokenEmpty      = "setting.gotify_token_empty"
+	MsgSettingGotifyUrlInvalid      = "setting.gotify_url_invalid"
+	MsgSettingTelegramBotTokenEmpty = "setting.telegram_bot_token_empty"
+	MsgSettingTelegramChatIdEmpty   = "setting.telegram_chat_id_empty"
+	MsgSettingTelegramChatIdInvalid = "setting.telegram_chat_id_invalid"
+	MsgSettingUrlMustHttp           = "setting.url_must_http"
+	MsgSettingSaved                 = "setting.saved"
 )
 
 // Deployment related messages (io.net)